rename synchronization analysis scripts
Rename the shiny new lnav wrapper to lnav-chef-sync-marks: we
mostly fixed drift issues, but it's still useful for debugging
general synchronization issues.

Meanwhile rename the old script back to its original name: in
case anyone still expects it to be there, ensure they can find
it and will see that it's obsolete.